Created authRouter(login, register, logout) and profileRouter, and moved secureApiRouter into its own file for organization
